% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Attachment_download}
\alias{Attachment_download}
\title{Download an attachment to a local path}
\usage{
Attachment_download(model_name, att_name, att_path)
}
\arguments{
\item{model_name}{model name}

\item{att_name}{attachment name}

\item{att_path}{path of the attachment (inluding file name)}
}
\value{
operation message or error message
}
\description{
Download an attachment to a local path
}
