% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Attachment_list}
\alias{Attachment_list}
\title{List all attchments of a model}
\usage{
Attachment_list(model_name)
}
\arguments{
\item{model_name}{name of the model to check for attachments}
}
\description{
List all attchments of a model
}
