% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Attachment_remove}
\alias{Attachment_remove}
\title{Remove an attachment from a model}
\usage{
Attachment_remove(model_name, att_name)
}
\arguments{
\item{model_name}{model name}

\item{att_name}{attachment name}
}
\value{
operation message or error message
}
\description{
Remove an attachment from a model
}
