% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Attachment_upload}
\alias{Attachment_upload}
\title{Upload an attqchment to a model from local path}
\usage{
Attachment_upload(model_name, att_name, att_path)
}
\arguments{
\item{model_name}{model name}

\item{att_name}{attachment name}

\item{att_path}{path of the attachment (inluding file name)}
}
\description{
Upload an attqchment to a model from local path
}
