% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Composer_config_set}
\alias{Composer_config_set}
\title{Set composer configuration from a local file}
\usage{
Composer_config_set(path)
}
\arguments{
\item{path}{path of the composer configuration file}
}
\description{
Set composer configuration from a local file
}
