% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Composer_config_show}
\alias{Composer_config_show}
\title{Show composer configuration}
\usage{
Composer_config_show()
}
\value{
instance info or error message
}
\description{
Show composer configuration
}
