% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Composer_create}
\alias{Composer_create}
\title{Create a composer from a local file}
\usage{
Composer_create(name, path)
}
\arguments{
\item{name}{composer name}

\item{path}{path of the composer file}
}
\description{
Create a composer from a local file
}
