% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Composer_delete}
\alias{Composer_delete}
\title{Delete a composer}
\usage{
Composer_delete(name)
}
\arguments{
\item{name}{name of the composer to delete}
}
\value{
operation message or error message
}
\description{
Delete a composer
}
