% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Composer_show}
\alias{Composer_show}
\title{Show content of a composer}
\usage{
Composer_show(name)
}
\arguments{
\item{name}{composer name}
}
\value{
instance info or error message
}
\description{
Show content of a composer
}
