% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Connect}
\alias{Connect}
\title{Connect to Fastscore, set proxy_prefix}
\usage{
Connect(proxy_prefix)
}
\arguments{
\item{proxy_prefix}{prefix of connect}
}
\description{
Connect to Fastscore, set proxy_prefix
}
