% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Engine_inspect}
\alias{Engine_inspect}
\title{Inspect an engine}
\usage{
Engine_inspect(name)
}
\arguments{
\item{name}{engine name}
}
\value{
instance info or error message
}
\description{
Inspect an engine
}
