% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Engine_pause}
\alias{Engine_pause}
\title{Pause an engine}
\usage{
Engine_pause(name)
}
\arguments{
\item{name}{engine name}
}
\value{
operation message or error message
}
\description{
Pause an engine
}
