% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Engine_reset}
\alias{Engine_reset}
\title{Reset an engine, or reset all engines}
\usage{
Engine_reset(name_or_all)
}
\arguments{
\item{name}{engine name, "all" to reset all engines}
}
\value{
operation message or error message
}
\description{
Reset an engine, or reset all engines
}
