% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Engine_unpause}
\alias{Engine_unpause}
\title{Unpause an engine}
\usage{
Engine_unpause(name)
}
\arguments{
\item{name}{engine name}
}
\value{
operation message or error message
}
\description{
Unpause an engine
}
