% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{import}
\alias{import}
\title{import a list of certain instances from a given directory}
\usage{
import(namelist, type, dir)
}
\arguments{
\item{namelist}{a list of instance names}

\item{type}{instance type ("model", "schema", "stream")}

\item{dir}{directory path (example: ".../fastscore/library/")}
}
\value{
operation message or error message
}
\description{
import a list of certain instances from a given directory
}
\examples{
import(list("array-double", "string"), "schema", "fastscore/library/")
}
