% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Model_add}
\alias{Model_add}
\title{Add a model from local file}
\usage{
Model_add(name, path)
}
\arguments{
\item{name}{model name}

\item{path}{model path (including file name)}
}
\description{
Add a model from local file
}
