% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Model_input}
\alias{Model_input}
\title{Send input to a (rest) engine slot}
\usage{
Model_input(engine_name, slot)
}
\arguments{
\item{engine_name}{engine name}

\item{slot}{engine slot to put input}
}
\value{
instance info or error message
}
\description{
Send input to a (rest) engine slot
}
