% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Model_inspect}
\alias{Model_inspect}
\title{Inspect a model}
\usage{
Model_inspect(name)
}
\arguments{
\item{name}{model name}
}
\value{
instance info or error message
}
\description{
Inspect a model
}
