% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Model_remove}
\alias{Model_remove}
\title{Remove a model}
\usage{
Model_remove(name)
}
\arguments{
\item{name}{model name}
}
\value{
operation message or error message
}
\description{
Remove a model
}
