% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Model_show}
\alias{Model_show}
\title{Show content of a model}
\usage{
Model_show(name)
}
\arguments{
\item{name}{model name}
}
\value{
instance info or error message
}
\description{
Show content of a model
}
