% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{remove_all}
\alias{remove_all}
\title{remove all model/stream/schema loaded}
\usage{
remove_all(type)
}
\arguments{
\item{type}{type of instance to clear ("model", "schema", "stream")}
}
\description{
remove all model/stream/schema loaded
}
\examples{
remove_all("model")
}
