% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Run}
\alias{Run}
\title{Run a model in an engine, need to specify input and output streams}
\usage{
Run(engine_name, model_name, stream0_name, stream1_name)
}
\arguments{
\item{engine_name}{engine name}

\item{model_name}{model name}

\item{stream0_name}{input stream name}

\item{stream1_name}{output stream name}
}
\value{
operation message or error message
}
\description{
Run a model in an engine, need to specify input and output streams
}
