% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Schema_add}
\alias{Schema_add}
\title{Add a schema from local path}
\usage{
Schema_add(name, path)
}
\arguments{
\item{name}{schema name}

\item{path}{path of file (including file name)}
}
\value{
operation message or error message
}
\description{
Add a schema from local path
}
