% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Schema_remove}
\alias{Schema_remove}
\title{Remove a schema}
\usage{
Schema_remove(name)
}
\arguments{
\item{name}{schema name}
}
\value{
operation message or error message
}
\description{
Remove a schema
}
