% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Sensor_add}
\alias{Sensor_add}
\title{Add a sensor from local path}
\usage{
Sensor_add(name, path)
}
\arguments{
\item{name}{sensor name}

\item{path}{sensor file path (including file name)}
}
\value{
operation message or error message
}
\description{
Add a sensor from local path
}
