% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Sensor_inspect}
\alias{Sensor_inspect}
\title{Inspect a sensor}
\usage{
Sensor_inspect(name)
}
\arguments{
\item{name}{sensor name}
}
\value{
instance info or error message
}
\description{
Inspect a sensor
}
