% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Sensor_install}
\alias{Sensor_install}
\title{Install a sensor}
\usage{
Sensor_install(name)
}
\arguments{
\item{name}{sensor name}
}
\value{
operation message or error message
}
\description{
Install a sensor
}
