% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Sensor_remove}
\alias{Sensor_remove}
\title{Remove a sensor}
\usage{
Sensor_remove(name)
}
\arguments{
\item{name}{sensor name}
}
\value{
operation message or error message
}
\description{
Remove a sensor
}
