% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Sensor_show}
\alias{Sensor_show}
\title{Show content of a sensor}
\usage{
Sensor_show(name)
}
\arguments{
\item{name}{schema name}
}
\value{
instance info or error message
}
\description{
Show content of a sensor
}
