% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Sensor_uninstall}
\alias{Sensor_uninstall}
\title{Uninstall a sensor}
\usage{
Sensor_uninstall(name)
}
\arguments{
\item{name}{sensor name}
}
\value{
operation message or error message
}
\description{
Uninstall a sensor
}
