% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Stream_add}
\alias{Stream_add}
\title{Add a stream from local file}
\usage{
Stream_add(name, path)
}
\arguments{
\item{name}{stream name}

\item{path}{stream file path (including file name)}
}
\value{
operation message or error message
}
\description{
Add a stream from local file
}
