% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Stream_attach}
\alias{Stream_attach}
\title{Attach a stream to an engine, specify slot}
\usage{
Stream_attach(slot, stream_name, engine_name)
}
\arguments{
\item{slot}{attach slot number}

\item{stream_name}{stream name}

\item{engine_name}{engine name}
}
\value{
operation message or error message
}
\description{
Attach a stream to an engine, specify slot
}
