% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Stream_inspect}
\alias{Stream_inspect}
\title{Inspect a stream}
\usage{
Stream_inspect(stream_name, engine_name)
}
\arguments{
\item{stream_name}{stream name}

\item{engine_name}{engine name}
}
\description{
Inspect a stream
}
