% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Stream_remove}
\alias{Stream_remove}
\title{Remove a stream}
\usage{
Stream_remove(name)
}
\arguments{
\item{name}{stream name}
}
\value{
operation message or error message
}
\description{
Remove a stream
}
