% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Stream_sample}
\alias{Stream_sample}
\title{Get samples from a stream, specify count}
\usage{
Stream_sample(stream_name, engine_name, count)
}
\arguments{
\item{stream_name}{stream name}

\item{engine_name}{engine name}

\item{count}{number of samples}
}
\value{
samples or error message
}
\description{
Get samples from a stream, specify count
}
