% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Stream_show}
\alias{Stream_show}
\title{Show content of a stream}
\usage{
Stream_show(name)
}
\arguments{
\item{name}{stream name}
}
\value{
instance info or error message
}
\description{
Show content of a stream
}
