% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Use}
\alias{Use}
\title{Use a certain engine}
\usage{
Use(engine_name)
}
\arguments{
\item{engine_name}{engine name}
}
\value{
operation message or error message
}
\description{
Use a certain engine
}
