% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotoRFastscore.R
\name{Config_set}
\alias{Config_set}
\title{Set configuration from a local file}
\usage{
Config_set(path)
}
\arguments{
\item{path}{path of the configuration file}
}
\description{
Set configuration from a local file
}
