% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastscoreRSDK.R
\name{Model_load}
\alias{Model_load}
\title{Load a model to an engine}
\usage{
Model_load(model_name, engine_name)
}
\arguments{
\item{model_name}{model name}

\item{engine_name}{model name}
}
\value{
operation message or error message
}
\description{
Load a model to an engine
}
